Implemented pausing and gamestates
The game now has three GameStates, Paused, in which it starts, Playing
and PostGame. The game starts on paused and waits one second to listen
for input in order to un-pause or return to menu. Upon pressing return
to menu the player is returned to the menu, duh. When in Playing mode
the player can press the Pause button to pause, or just play the game