Almost finished, found a bug, don't wanna do math.
Finished commenting all of my code so far. Found a bug that makes
the ball dissappear. I don't want to do the math to figure out
which side of a brick the ball is hitting so that I'm not just
sending the ball down whenever it is hits a ball, even from the top